HBX-1586: Move classes 'MetaDataDialect' and 'MetaDataDialectFactory' from package 'org.hibernate.tool.api.reveng' to 'org.hibernate.tool.api.dialect'
