+++ main/src/main/java/org/hibernate/cfg/JDBCReaderFactory.java
@@ -7,8 +7,8 @@ import org.hibernate.engine.jdbc.connections.spi.ConnectionProvider;
 import org.hibernate.engine.jdbc.spi.JdbcServices;
 import org.hibernate.exception.spi.SQLExceptionConverter;
 import org.hibernate.service.ServiceRegistry;
-import org.hibernate.tool.api.reveng.MetaDataDialect;
-import org.hibernate.tool.api.reveng.MetaDataDialectFactory;
+import org.hibernate.tool.api.dialect.MetaDataDialect;
+import org.hibernate.tool.api.dialect.MetaDataDialectFactory;
 import org.hibernate.tool.api.reveng.ReverseEngineeringStrategy;
 
 final public class JDBCReaderFactory {
