+++ main/src/main/java/org/hibernate/cfg/reveng/BasicColumnProcessor.java
@@ -8,7 +8,7 @@ import org.hibernate.JDBCException;
 import org.hibernate.cfg.JDBCBinderException;
 import org.hibernate.mapping.Column;
 import org.hibernate.mapping.Table;
-import org.hibernate.tool.api.reveng.MetaDataDialect;
+import org.hibernate.tool.api.dialect.MetaDataDialect;
 import org.hibernate.tool.api.reveng.ProgressListener;
 import org.hibernate.tool.api.reveng.ReverseEngineeringStrategy;
 import org.hibernate.tool.api.reveng.TableIdentifier;
