+++ main/src/main/java/org/hibernate/cfg/reveng/DefaultDatabaseCollector.java
@@ -9,7 +9,7 @@ import java.util.Map.Entry;
 
 import org.hibernate.internal.util.StringHelper;
 import org.hibernate.mapping.Table;
-import org.hibernate.tool.api.reveng.MetaDataDialect;
+import org.hibernate.tool.api.dialect.MetaDataDialect;
 import org.hibernate.tool.util.TableNameQualifier;
 
 public class DefaultDatabaseCollector extends AbstractDatabaseCollector  {
