+++ test/common/src/main/java/org/hibernate/tool/cfg/DriverMetaData/TestCase.java
@@ -15,8 +15,8 @@ import org.hibernate.cfg.reveng.DefaultDatabaseCollector;
 import org.hibernate.engine.jdbc.connections.spi.ConnectionProvider;
 import org.hibernate.engine.jdbc.spi.JdbcServices;
 import org.hibernate.service.ServiceRegistry;
-import org.hibernate.tool.api.reveng.MetaDataDialect;
-import org.hibernate.tool.api.reveng.MetaDataDialectFactory;
+import org.hibernate.tool.api.dialect.MetaDataDialect;
+import org.hibernate.tool.api.dialect.MetaDataDialectFactory;
 import org.hibernate.tool.api.reveng.ReverseEngineeringRuntimeInfo;
 import org.hibernate.tool.internal.dialect.JDBCMetaDataDialect;
 import org.hibernate.tools.test.util.JUnitUtil;
