+++ test/oracle/src/main/java/org/hibernate/cfg/reveng/dialect/TestCase.java
@@ -7,7 +7,7 @@ import org.hibernate.cfg.reveng.DefaultDatabaseCollector;
 import org.hibernate.engine.jdbc.connections.spi.ConnectionProvider;
 import org.hibernate.engine.jdbc.spi.JdbcServices;
 import org.hibernate.service.ServiceRegistry;
-import org.hibernate.tool.api.reveng.MetaDataDialect;
+import org.hibernate.tool.api.dialect.MetaDataDialect;
 import org.hibernate.tool.api.reveng.ReverseEngineeringRuntimeInfo;
 import org.hibernate.tool.internal.dialect.OracleMetaDataDialect;
 import org.hibernate.tools.test.util.JdbcUtil;
